feat: timestampstatus proxy adminPort endpoint refactored to grpc
Refs: XRDDEV-2673
