+++ src/proxy/proto-model/build.gradle
@@ -15,6 +15,7 @@ sourceSets {
 
 dependencies {
     implementation (project(':common:common-core'))
+    implementation (project(':common:common-globalconf')) //for the ee.ria.xroad.common.DiagnosticsStatus
     implementation (project(':common:common-rpc'))
 //    implementation (libs.slf4j.api)
